[rtree] Factor out rtree base features
These will be moved into thin-provisioning-tools
